# Optimized store function by removing repetitions