Recently, I did some numerical experiments with many arms more than 5000 arms, and horizon 100000.

I found that it is extremely slow to repeat using 'for' statement with thousands of arms, in `store` function of SMPyBandits/Environment/Evaluator.py.
Therefore, I optimized that `store` function by removing repetitions.

The optimized code consumes about only half the execution time of the original code!
I compared running times of original code and my optimized code below.


nbArms = 5000, horizon = 50000
original: 124s, optimized: 66s
```
>>> import timeit
>>>
>>> timeit.timeit('''
... import numpy as np
...
... nbArms = 5000
... horizon = 50000
... choices = np.random.randint(0, nbArms, horizon)
...
... np.array([1 * (choices == armId) for armId in range(nbArms)])
... ''', number=50)
124.48019195400002


>>> import timeit
>>>
>>> timeit.timeit('''
... import numpy as np
...
... nbArms = 5000
... horizon = 50000
... choices = np.random.randint(0, nbArms, horizon)
... select_rows = choices.reshape(1, -1)
...
... np.transpose(np.identity(nbArms, dtype=int)[select_rows][0])
... ''', number=50)
65.51110079
```

nbArms = 1000, horizon = 50000
original: 25s, optimized: 10s

nbArms = 5000, horizon = 50000
original: 26s, optimized: 16s

nbArms = 5000, horizon = 50000
original: 0.36s, optimized: 0.04s

Test environment:
MacBook Pro (15-inch, 2017)
2.8 GHz Quad Core Intel Core i7
16GB 2133 MHz LPDDR3